internal: Move packages to top level
These packages are useful in themselves, so they should be importable.
