+++ api/metrics/v1/http.go
@@ -11,7 +11,7 @@ import (
 	"github.com/go-kit/kit/log"
 	"github.com/prometheus/client_golang/prometheus"
 
-	"github.com/observatorium/observatorium/internal/proxy"
+	"github.com/observatorium/observatorium/proxy"
 )
 
 const (
